perf: dynamically import time saved badge updater
